penny/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
httpx==0.27.2
python-multipart==0.0.9
pydantic==2.8.2
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
asyncpg==0.29.0
structlog==24.4.0
